+++ apps/contribute/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   New page for editing tags, incorporated into the flow.
+
 ### Changed
 
 -   Upgraded `@aws-sdk` dependencies.